Python/pyproject.toml
[tool.poetry]
name = "inat_project_extractor"
version = "0.1.0"
description = "For exporting iNaturalist project data"
authors = ["Stephen Fuqua <stephen@safnet.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.1"
requests = "^2.25.1"
ConfigArgParse = "^1.2.3"
python-dotenv = "^0.15.0"
opnieuw = "^1.1.0"
errorhandler = "^2.0.1"
pandas = "^1.2.0"
numpy = "1.19.3"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
